refactor: move staticUdf function
