+++ dinky-admin/src/main/java/org/dinky/controller/JarController.java
@@ -86,7 +86,7 @@ public class JarController {
     @GetMapping("/udf/geUdfs")
     @ApiOperation("Get UDFs")
     public Result<List<CascaderVO>> getUdfs() {
-        List<UDF> staticUdfs = UDFUtils.getStaticUdfs();
+        List<UDF> staticUdfs = UDFUtil.getStaticUdfs();
         List<UDF> dynamicUdfs =
                 taskService.getAllUdfEnabled().stream().map(UDFUtils::taskToUDF).collect(Collectors.toList());
         List<UDF> allUdfs = new ArrayList<>(staticUdfs);
